+++ src/cdef_tmpl_8.rs
@@ -318,16 +318,16 @@ unsafe extern "C" fn padding(
     edges: CdefEdgeFlags,
 ) {
     let mut x_start: libc::c_int = -(2 as libc::c_int);
-    let mut x_end: libc::c_int = w + 2 as libc::c_int;
+    let mut x_end: libc::c_int = w + 2;
     let mut y_start: libc::c_int = -(2 as libc::c_int);
-    let mut y_end: libc::c_int = h + 2 as libc::c_int;
+    let mut y_end: libc::c_int = h + 2;
     if edges as libc::c_uint & CDEF_HAVE_TOP as libc::c_int as libc::c_uint == 0 {
         fill(
             tmp
                 .offset(-(2 as libc::c_int as isize))
                 .offset(-((2 * tmp_stride) as isize)),
             tmp_stride,
-            w + 4 as libc::c_int,
+            w + 4,
             2 as libc::c_int,
         );
         y_start = 0 as libc::c_int;
@@ -338,7 +338,7 @@ unsafe extern "C" fn padding(
                 .offset((h as isize * tmp_stride) as isize)
                 .offset(-(2 as libc::c_int as isize)),
             tmp_stride,
-            w + 4 as libc::c_int,
+            w + 4,
             2 as libc::c_int,
         );
         y_end -= 2 as libc::c_int;
@@ -366,7 +366,7 @@ unsafe extern "C" fn padding(
         x_end -= 2 as libc::c_int;
     }
     let mut y: libc::c_int = y_start;
-    while y < 0 as libc::c_int {
+    while y < 0 {
         let mut x: libc::c_int = x_start;
         while x < x_end {
             *tmp
@@ -381,7 +381,7 @@ unsafe extern "C" fn padding(
     let mut y_0 = 0;
     while y_0 < h {
         let mut x_0: libc::c_int = x_start;
-        while x_0 < 0 as libc::c_int {
+        while x_0 < 0 {
             *tmp
                 .offset(
                     (x_0 as isize + y_0 as isize * tmp_stride) as isize,
@@ -430,8 +430,8 @@ unsafe extern "C" fn cdef_filter_block_c(
     edges: CdefEdgeFlags,
 ) {
     let tmp_stride: ptrdiff_t = 12 as libc::c_int as ptrdiff_t;
-    if !((w == 4 as libc::c_int || w == 8 as libc::c_int)
-        && (h == 4 as libc::c_int || h == 8 as libc::c_int))
+    if !((w == 4 || w == 8)
+        && (h == 4 || h == 8))
     {
         unreachable!();
     }
@@ -442,9 +442,9 @@ unsafe extern "C" fn cdef_filter_block_c(
         .offset(2);
     padding(tmp, tmp_stride, dst, dst_stride, left, top, bottom, w, h, edges);
     if pri_strength != 0 {
-        let bitdepth_min_8: libc::c_int = 8 as libc::c_int - 8 as libc::c_int;
+        let bitdepth_min_8: libc::c_int = 8 as libc::c_int - 8;
         let pri_tap: libc::c_int = 4 as libc::c_int
-            - (pri_strength >> bitdepth_min_8 & 1 as libc::c_int);
+            - (pri_strength >> bitdepth_min_8 & 1);
         let pri_shift: libc::c_int = imax(
             0 as libc::c_int,
             damping - ulog2(pri_strength as libc::c_uint),
@@ -460,16 +460,16 @@ unsafe extern "C" fn cdef_filter_block_c(
                     let mut min: libc::c_int = px;
                     let mut pri_tap_k: libc::c_int = pri_tap;
                     let mut k = 0;
-                    while k < 2 as libc::c_int {
+                    while k < 2 {
                         let off1: libc::c_int = dav1d_cdef_directions[(dir
-                            + 2 as libc::c_int) as usize][k as usize] as libc::c_int;
+                            + 2) as usize][k as usize] as libc::c_int;
                         let p0: libc::c_int = *tmp.offset((x + off1) as isize)
                             as libc::c_int;
                         let p1: libc::c_int = *tmp.offset((x - off1) as isize)
                             as libc::c_int;
                         sum += pri_tap_k * constrain(p0 - px, pri_strength, pri_shift);
                         sum += pri_tap_k * constrain(p1 - px, pri_strength, pri_shift);
-                        pri_tap_k = pri_tap_k & 3 as libc::c_int | 2 as libc::c_int;
+                        pri_tap_k = pri_tap_k & 3 | 2;
                         min = umin(p0 as libc::c_uint, min as libc::c_uint)
                             as libc::c_int;
                         max = imax(p0, max);
@@ -477,9 +477,9 @@ unsafe extern "C" fn cdef_filter_block_c(
                             as libc::c_int;
                         max = imax(p1, max);
                         let off2: libc::c_int = dav1d_cdef_directions[(dir
-                            + 4 as libc::c_int) as usize][k as usize] as libc::c_int;
+                            + 4) as usize][k as usize] as libc::c_int;
                         let off3: libc::c_int = dav1d_cdef_directions[(dir
-                            + 0 as libc::c_int) as usize][k as usize] as libc::c_int;
+                            + 0) as usize][k as usize] as libc::c_int;
                         let s0: libc::c_int = *tmp.offset((x + off2) as isize)
                             as libc::c_int;
                         let s1: libc::c_int = *tmp.offset((x - off2) as isize)
@@ -512,8 +512,8 @@ unsafe extern "C" fn cdef_filter_block_c(
                             x as isize,
                         ) = iclip(
                         px
-                            + (sum - (sum < 0 as libc::c_int) as libc::c_int
-                                + 8 as libc::c_int >> 4 as libc::c_int),
+                            + (sum - (sum < 0) as libc::c_int
+                                + 8 >> 4),
                         min,
                         max,
                     ) as pixel;
@@ -534,9 +534,9 @@ unsafe extern "C" fn cdef_filter_block_c(
                     let mut sum_0 = 0;
                     let mut pri_tap_k_0: libc::c_int = pri_tap;
                     let mut k_0 = 0;
-                    while k_0 < 2 as libc::c_int {
+                    while k_0 < 2 {
                         let off: libc::c_int = dav1d_cdef_directions[(dir
-                            + 2 as libc::c_int) as usize][k_0 as usize] as libc::c_int;
+                            + 2) as usize][k_0 as usize] as libc::c_int;
                         let p0_0: libc::c_int = *tmp.offset((x_0 + off) as isize)
                             as libc::c_int;
                         let p1_0: libc::c_int = *tmp.offset((x_0 - off) as isize)
@@ -547,15 +547,15 @@ unsafe extern "C" fn cdef_filter_block_c(
                         sum_0
                             += pri_tap_k_0
                                 * constrain(p1_0 - px_0, pri_strength, pri_shift);
-                        pri_tap_k_0 = pri_tap_k_0 & 3 as libc::c_int | 2 as libc::c_int;
+                        pri_tap_k_0 = pri_tap_k_0 & 3 | 2;
                         k_0 += 1;
                     }
                     *dst
                         .offset(
                             x_0 as isize,
                         ) = (px_0
-                        + (sum_0 - (sum_0 < 0 as libc::c_int) as libc::c_int
-                            + 8 as libc::c_int >> 4 as libc::c_int)) as pixel;
+                        + (sum_0 - (sum_0 < 0) as libc::c_int
+                            + 8 >> 4)) as pixel;
                     x_0 += 1;
                 }
                 dst = dst.offset(dst_stride as isize);
@@ -577,11 +577,11 @@ unsafe extern "C" fn cdef_filter_block_c(
                 let px_1: libc::c_int = *dst.offset(x_1 as isize) as libc::c_int;
                 let mut sum_1 = 0;
                 let mut k_1 = 0;
-                while k_1 < 2 as libc::c_int {
+                while k_1 < 2 {
                     let off1_0: libc::c_int = dav1d_cdef_directions[(dir
-                        + 4 as libc::c_int) as usize][k_1 as usize] as libc::c_int;
+                        + 4) as usize][k_1 as usize] as libc::c_int;
                     let off2_0: libc::c_int = dav1d_cdef_directions[(dir
-                        + 0 as libc::c_int) as usize][k_1 as usize] as libc::c_int;
+                        + 0) as usize][k_1 as usize] as libc::c_int;
                     let s0_0: libc::c_int = *tmp.offset((x_1 + off1_0) as isize)
                         as libc::c_int;
                     let s1_0: libc::c_int = *tmp.offset((x_1 - off1_0) as isize)
@@ -605,8 +605,8 @@ unsafe extern "C" fn cdef_filter_block_c(
                     .offset(
                         x_1 as isize,
                     ) = (px_1
-                    + (sum_1 - (sum_1 < 0 as libc::c_int) as libc::c_int
-                        + 8 as libc::c_int >> 4 as libc::c_int)) as pixel;
+                    + (sum_1 - (sum_1 < 0) as libc::c_int
+                        + 8 >> 4)) as pixel;
                 x_1 += 1;
             }
             dst = dst.offset(dst_stride as isize);
@@ -704,7 +704,7 @@ unsafe extern "C" fn cdef_find_dir_c(
     stride: ptrdiff_t,
     var: *mut libc::c_uint,
 ) -> libc::c_int {
-    let bitdepth_min_8: libc::c_int = 8 as libc::c_int - 8 as libc::c_int;
+    let bitdepth_min_8: libc::c_int = 8 as libc::c_int - 8;
     let mut partial_sum_hv: [[libc::c_int; 8]; 2] = [
         [0 as libc::c_int, 0, 0, 0, 0, 0, 0, 0],
         [0; 8],
@@ -720,26 +720,26 @@ unsafe extern "C" fn cdef_find_dir_c(
         [0; 11],
     ];
     let mut y = 0;
-    while y < 8 as libc::c_int {
+    while y < 8 {
         let mut x = 0;
-        while x < 8 as libc::c_int {
+        while x < 8 {
             let px: libc::c_int = (*img.offset(x as isize) as libc::c_int
-                >> bitdepth_min_8) - 128 as libc::c_int;
+                >> bitdepth_min_8) - 128;
             partial_sum_diag[0][(y + x) as usize] += px;
             partial_sum_alt[0 as libc::c_int
-                as usize][(y + (x >> 1 as libc::c_int)) as usize] += px;
+                as usize][(y + (x >> 1)) as usize] += px;
             partial_sum_hv[0][y as usize] += px;
             partial_sum_alt[1 as libc::c_int
-                as usize][(3 as libc::c_int + y - (x >> 1 as libc::c_int)) as usize]
+                as usize][(3 as libc::c_int + y - (x >> 1)) as usize]
                 += px;
             partial_sum_diag[1 as libc::c_int
                 as usize][(7 as libc::c_int + y - x) as usize] += px;
             partial_sum_alt[2 as libc::c_int
-                as usize][(3 as libc::c_int - (y >> 1 as libc::c_int) + x) as usize]
+                as usize][(3 as libc::c_int - (y >> 1) + x) as usize]
                 += px;
             partial_sum_hv[1][x as usize] += px;
             partial_sum_alt[3 as libc::c_int
-                as usize][((y >> 1 as libc::c_int) + x) as usize] += px;
+                as usize][((y >> 1) + x) as usize] += px;
             x += 1;
         }
         img = img.offset(stride as isize);
@@ -756,7 +756,7 @@ unsafe extern "C" fn cdef_find_dir_c(
         0,
     ];
     let mut n = 0;
-    while n < 8 as libc::c_int {
+    while n < 8 {
         cost[2 as libc::c_int
             as usize] = (cost[2])
             .wrapping_add(
@@ -789,7 +789,7 @@ unsafe extern "C" fn cdef_find_dir_c(
         120 as libc::c_int as uint16_t,
     ];
     let mut n_0 = 0;
-    while n_0 < 7 as libc::c_int {
+    while n_0 < 7 {
         let d: libc::c_int = div_table[n_0 as usize] as libc::c_int;
         cost[0 as libc::c_int
             as usize] = (cost[0])
@@ -820,23 +820,23 @@ unsafe extern "C" fn cdef_find_dir_c(
         .wrapping_add(
             (partial_sum_diag[0][7]
                 * partial_sum_diag[0][7]
-                * 105 as libc::c_int) as libc::c_uint,
+                * 105) as libc::c_uint,
         );
     cost[4 as libc::c_int
         as usize] = (cost[4])
         .wrapping_add(
             (partial_sum_diag[1][7]
                 * partial_sum_diag[1][7]
-                * 105 as libc::c_int) as libc::c_uint,
+                * 105) as libc::c_uint,
         );
     let mut n_1 = 0;
-    while n_1 < 4 as libc::c_int {
+    while n_1 < 4 {
         let cost_ptr: *mut libc::c_uint = &mut *cost
             .as_mut_ptr()
-            .offset((n_1 * 2 as libc::c_int + 1 as libc::c_int) as isize)
+            .offset((n_1 * 2 + 1) as isize)
             as *mut libc::c_uint;
         let mut m = 0;
-        while m < 5 as libc::c_int {
+        while m < 5 {
             *cost_ptr = (*cost_ptr)
                 .wrapping_add(
                     (partial_sum_alt[n_1 as usize][(3 as libc::c_int + m) as usize]
@@ -847,8 +847,8 @@ unsafe extern "C" fn cdef_find_dir_c(
         }
         *cost_ptr = (*cost_ptr).wrapping_mul(105 as libc::c_int as libc::c_uint);
         let mut m_0 = 0;
-        while m_0 < 3 as libc::c_int {
-            let d_0: libc::c_int = div_table[(2 as libc::c_int * m_0 + 1 as libc::c_int)
+        while m_0 < 3 {
+            let d_0: libc::c_int = div_table[(2 as libc::c_int * m_0 + 1)
                 as usize] as libc::c_int;
             *cost_ptr = (*cost_ptr)
                 .wrapping_add(
@@ -867,7 +867,7 @@ unsafe extern "C" fn cdef_find_dir_c(
     let mut best_dir = 0;
     let mut best_cost: libc::c_uint = cost[0];
     let mut n_2 = 1;
-    while n_2 < 8 as libc::c_int {
+    while n_2 < 8 {
         if cost[n_2 as usize] > best_cost {
             best_cost = cost[n_2 as usize];
             best_dir = n_2;
@@ -875,7 +875,7 @@ unsafe extern "C" fn cdef_find_dir_c(
         n_2 += 1;
     }
     *var = best_cost.wrapping_sub(cost[(best_dir ^ 4 as libc::c_int) as usize])
-        >> 10 as libc::c_int;
+        >> 10;
     return best_dir;
 }
