+++ src/intra_edge.rs
@@ -297,7 +297,7 @@ unsafe extern "C" fn init_mode_node(
     );
     if bl as libc::c_uint == BL_16X16 as libc::c_int as libc::c_uint {
         let mut n = 0;
-        while n < 4 as libc::c_int {
+        while n < 4 {
             let fresh0 = (*mem).nt;
             (*mem).nt = ((*mem).nt).offset(1);
             let nt: *mut EdgeTip = fresh0;
@@ -306,7 +306,7 @@ unsafe extern "C" fn init_mode_node(
                 &mut (*nt).node,
                 (bl as libc::c_uint).wrapping_add(1 as libc::c_int as libc::c_uint)
                     as BlockLevel,
-                ((if n == 3 as libc::c_int || n == 1 as libc::c_int && top_has_right == 0
+                ((if n == 3 || n == 1 && top_has_right == 0
                 {
                     0 as libc::c_int
                 } else {
@@ -314,8 +314,8 @@ unsafe extern "C" fn init_mode_node(
                         | EDGE_I422_TOP_HAS_RIGHT as libc::c_int
                         | EDGE_I420_TOP_HAS_RIGHT as libc::c_int
                 })
-                    | (if !(n == 0 as libc::c_int
-                        || n == 2 as libc::c_int && left_has_bottom != 0)
+                    | (if !(n == 0
+                        || n == 2 && left_has_bottom != 0)
                     {
                         0 as libc::c_int
                     } else {
@@ -328,7 +328,7 @@ unsafe extern "C" fn init_mode_node(
         }
     } else {
         let mut n_0 = 0;
-        while n_0 < 4 as libc::c_int {
+        while n_0 < 4 {
             let fresh1 = (*mem).nwc[bl as usize];
             (*mem).nwc[bl as usize] = ((*mem).nwc[bl as usize]).offset(1);
             let nwc_child: *mut EdgeBranch = fresh1;
@@ -338,10 +338,10 @@ unsafe extern "C" fn init_mode_node(
                 (bl as libc::c_uint).wrapping_add(1 as libc::c_int as libc::c_uint)
                     as BlockLevel,
                 mem,
-                !(n_0 == 3 as libc::c_int
-                    || n_0 == 1 as libc::c_int && top_has_right == 0) as libc::c_int,
-                (n_0 == 0 as libc::c_int
-                    || n_0 == 2 as libc::c_int && left_has_bottom != 0) as libc::c_int,
+                !(n_0 == 3
+                    || n_0 == 1 && top_has_right == 0) as libc::c_int,
+                (n_0 == 0
+                    || n_0 == 2 && left_has_bottom != 0) as libc::c_int,
             );
             n_0 += 1;
         }
@@ -365,16 +365,16 @@ pub unsafe extern "C" fn dav1d_init_mode_tree(
             as usize] = &mut *root.offset(1) as *mut EdgeBranch;
         mem
             .nwc[BL_64X64 as libc::c_int
-            as usize] = &mut *root.offset((1 as libc::c_int + 4 as libc::c_int) as isize)
+            as usize] = &mut *root.offset((1 as libc::c_int + 4) as isize)
             as *mut EdgeBranch;
         mem
             .nwc[BL_32X32 as libc::c_int
             as usize] = &mut *root
-            .offset((1 as libc::c_int + 4 as libc::c_int + 16 as libc::c_int) as isize)
+            .offset((1 as libc::c_int + 4 + 16) as isize)
             as *mut EdgeBranch;
         init_mode_node(root, BL_128X128, &mut mem, 1 as libc::c_int, 0 as libc::c_int);
         if !(mem.nwc[BL_128X128 as libc::c_int as usize]
-            == &mut *root.offset((1 as libc::c_int + 4 as libc::c_int) as isize)
+            == &mut *root.offset((1 as libc::c_int + 4) as isize)
                 as *mut EdgeBranch)
         {
             unreachable!();
@@ -382,7 +382,7 @@ pub unsafe extern "C" fn dav1d_init_mode_tree(
         if !(mem.nwc[BL_64X64 as libc::c_int as usize]
             == &mut *root
                 .offset(
-                    (1 as libc::c_int + 4 as libc::c_int + 16 as libc::c_int) as isize,
+                    (1 as libc::c_int + 4 + 16) as isize,
                 ) as *mut EdgeBranch)
         {
             unreachable!();
@@ -390,8 +390,8 @@ pub unsafe extern "C" fn dav1d_init_mode_tree(
         if !(mem.nwc[BL_32X32 as libc::c_int as usize]
             == &mut *root
                 .offset(
-                    (1 as libc::c_int + 4 as libc::c_int + 16 as libc::c_int
-                        + 64 as libc::c_int) as isize,
+                    (1 as libc::c_int + 4 + 16
+                        + 64) as isize,
                 ) as *mut EdgeBranch)
         {
             unreachable!();
@@ -406,11 +406,11 @@ pub unsafe extern "C" fn dav1d_init_mode_tree(
             as usize] = &mut *root.offset(1) as *mut EdgeBranch;
         mem
             .nwc[BL_32X32 as libc::c_int
-            as usize] = &mut *root.offset((1 as libc::c_int + 4 as libc::c_int) as isize)
+            as usize] = &mut *root.offset((1 as libc::c_int + 4) as isize)
             as *mut EdgeBranch;
         init_mode_node(root, BL_64X64, &mut mem, 1 as libc::c_int, 0 as libc::c_int);
         if !(mem.nwc[BL_64X64 as libc::c_int as usize]
-            == &mut *root.offset((1 as libc::c_int + 4 as libc::c_int) as isize)
+            == &mut *root.offset((1 as libc::c_int + 4) as isize)
                 as *mut EdgeBranch)
         {
             unreachable!();
@@ -418,7 +418,7 @@ pub unsafe extern "C" fn dav1d_init_mode_tree(
         if !(mem.nwc[BL_32X32 as libc::c_int as usize]
             == &mut *root
                 .offset(
-                    (1 as libc::c_int + 4 as libc::c_int + 16 as libc::c_int) as isize,
+                    (1 as libc::c_int + 4 + 16) as isize,
                 ) as *mut EdgeBranch)
         {
             unreachable!();
